backend/requirements.txt
Flask
Flask-Cors
psycopg
psycopg-binary
python-dotenv
